LayoutTests/platform/gtk/accessibility/aria-help-expected.txt

button
This tests that aria-help attribute works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL button.helpText should be AXHelp: click here (of type string). Was null (of type object).
PASS successfullyParsed is true

TEST COMPLETE
